locust-load-testing/requirements.txt
locust
PyYAML
Faker